DATA-226/requirements.txt
apache-airflow
requests
snowflake-connector-python
python-dotenv
